% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_hmdb_ms.R
\name{query_hmdb_ms}
\alias{query_hmdb_ms}
\title{Query HMDB MS Spectra and Retrieve Compound Metadata}
\usage{
query_hmdb_ms(
  mz_rt = NULL,
  ion_mode = "Positive",
  filter_adduct_type = NULL,
  tolerance = 5,
  tolerance_units = "ppm",
  ccs_predictors = NULL,
  ccs_tolerance = NULL
)
}
\arguments{
\item{mz_rt}{Character string or vector. Feature m/z and RT in the format
\code{"m/z@RT"}. Defaults to \code{NULL}. Can accept multiple values.}

\item{ion_mode}{Character string. MS ionization mode. One of
\code{"Positive"}, \code{"Negative"}, or \code{"Neutral"}. Default = \code{"Positive"}.}

\item{filter_adduct_type}{Character vector. Specific adducts to retain.
Defaults to \code{NULL}. If provided, results are filtered accordingly.}

\item{tolerance}{Numeric. Mass tolerance value.
Default = \code{5}.}

\item{tolerance_units}{Character string. Mass tolerance unit.
One of \code{"Da"} or \code{"ppm"}. Default = \code{"ppm"}.}

\item{ccs_predictors}{Character string or vector. CCS prediction model(s).
Examples: \code{"AllCCS"}, \code{"DarkChem"}, \code{"DeepCCS"}. Default = \code{NULL}.}

\item{ccs_tolerance}{Numeric. CCS tolerance value. Choose between c(1, 3, 5, 10). Default = \code{NULL}.}
}
\value{
A list with two data frames:
\itemize{
\item \code{compound_metadata}: Compound details scraped from individual
HMDB metabolite pages (name, taxonomy, framework, etc.).
\item \code{results_table_excluded}: Table of results where adducts
do not match the provided filter list.
}
}
\description{
This function queries the Human Metabolome Database (HMDB) Mass Spectrometry
(MS) search interface (specifically at 'LC-MS Search') using provided m/z and
retention time (RT), ion mode, adduct filters, tolerance levels,
and optional collision cross section (CCS) parameters.
It returns both detailed compound metadata from individual HMDB metabolite
pages and the excluded results table (those adducts not in the filter list).
}
\details{
The function submits form data to the HMDB MS Spectra Search interface,
retrieves the results table, filters adducts (if specified), and then
scrapes additional information for each compound from its dedicated page.
}
\examples{
\dontrun{
results <- query_hmdb_ms(
  mz_rt = "515.2597@10.7733",
  ion_mode = "Positive",
  filter_adduct_type = c("M+H", "M+Na"),
  tolerance = 5,
  tolerance_units = "ppm"
)

head(results$compound_metadata)
head(results$results_table_excluded)
}

}
\seealso{
The HMDB website \url{https://www.hmdb.ca/spectra/ms/search} and \url{https://www.docker.com/products/docker-desktop/} to download the Docker Desktop to be able to use this function.
}
\author{
John Lennon L. Calorio
}
